requirements.txt
openai>=1.0.0
flask>=3.0.0
flask-cors>=6.0.0
requests>=2.31.0
pyyaml>=6.0
